requirements.txt
requests==2.31.0            # GitHub API calls
python-dotenv==1.0.0        # Environment variables (for API tokens)
pandas==2.0.3               # CSV handling (optional but useful)
pytest==7.4.0               # Testing
tqdm==4.66.1                # Progress bars (for logging)